Create cheeseburger.html and link with index.html
Create cheeseburger.html and code boilerplate HTML. After basic
formatting, create a link to cheeseburger.html in index.html. To
ensure a way to return to previous page, create a link in index.html
to cheeseburger.html.

This is to create a basic web of pages for the Recipes website.